484 incorporate enumerations (#619)
* add monai/utils/enums.py

contains the str subsets

* incorporate enumeration NumpyPadMode

* incorporate enumeration GridSampleMode

test case removed since __init__ now requires a valid
mode and padding_mode even if __call__ overrides them

* incorporate enumeration InterpolateMode

* incorporate enumeration UpsampleMode

* incorporate enumeration BlendMode

* incorporate enumeration PytorchPadMode

* incorporate enumeration GridSamplePadMode

* incorporate enumeration Average

* incorporate enumeration MetricReduction

* incorporate enumeration LossReduction

* incorporate enumeration Weight

* incorporate enumeration Normalisation

* incorporate enumeration Activation

* incorporate enumeration ChannelMatching

* incorporate enumeration Method

* incorporate union GridSampleMode InterpolateMode

"nearest" left as default since it is valid for both enums

* include enumerations in tests/

any enum function which uses a default value uses an enum

* update type hinting of enum sequences

Sequence[object] was to allow a mixed sequence of enum and str
to pass the type check. However this would be a rare occurrence
and would have no reason for being. Of higher concern it allows
any sequenced subclass of object to pass the type check.

* add explicit enum type hint to instance variables