Make the flexible unet more extensible to different backbones (#5038)
Signed-off-by: binliu <binliu@nvidia.com>

Fixes #5011  .

### Description
Currently, the FlexibleUNet structure only supports the efficient-net
series as the backbone and it is hard to extend to other network
structures. A more extensible and convenient way to add more benchmark
or user-defined backbones will make it more flexible. I plan to make it
by doing steps below

1. A base backbone/encoder class that defines interfaces like the number
of output feature maps, a list of output feature map channels, string
names of backbones and so on.

2. A register that can dynamically register backbones to a dict that will
be used during the flexible unet initialization.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

Signed-off-by: binliu <binliu@nvidia.com>