Classification post-process for HoVerNet (#5462)
Signed-off-by: KumoLiu <yunl@nvidia.com>

Fixes #5180.

### Description
For the post-process in HoVerNet, now we have PR #5173 to process the
segmentation output(which has been merged) and PR #5186 to process the
classification output(which is under review now). We want a whole
transform containing these two components which can be easily used in
the MONAI model zoo bundle. Here has a draft version of how to combine
these components, if anyone has any suggestions, please feel free to
take it out.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

Signed-off-by: KumoLiu <yunl@nvidia.com>
Signed-off-by: yunliu <yunl@nvidia.com>
Co-authored-by: Behrooz Hashemian <3968947+drbeh@users.noreply.github.com>