CheckLabelShaped to check and correct label shapes (#5606)
adds CheckLabelShaped transform under monai.apps.auto3dseg:

- to check if image shape and label shape are the same. Currently, we
usually don't check for it, which results in potentially a
non-descriptive crash later, when computing loss or metric.
- to correct label shape if it is very similar to image shape.
Unfortunately some datasets, such as Hecktor22, has some labels (masks)
with a slightly different shapes (difference in only 1 voxel). And
because of that a user will have to either re-save the whole dataset, or
account for it during the training loop (which is inconvenient). this
utility auto corrects it.

We also already have this auto-correction in DataAnalyzer, and with this
transform, DataAnalyzer logic is simplified.

Signed-off-by: myron <amyronenko@nvidia.com>
Signed-off-by: Wenqi Li <wenqil@nvidia.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wenqi Li <831580+wyli@users.noreply.github.com>
Co-authored-by: Wenqi Li <wenqil@nvidia.com>