Fix SSIMLoss when batchsize bigger than 2 (#5908)
Signed-off-by: Jakob Weigand <jakob.weigand@tum.de>

Fixes #5886.

### Description

When SSIMLoss is called with a tensor of batchsize > 1, SSIMMetric is
called for each batch item individually. The result of each call to
SSIMMetric is stored in ssim_val. Subsequently, ssim_val is concatenated
to ssim_value, which is a collection of all calls to SSIMMetric. Before
the concatenation both ssim_val and ssim_value are given a new view of
size 1. As ssim_value collects the results of all earlier calls to
SSIMMetric, .view(1) raises an error for the second call (as ssim_value
has a different absolute size). This commit changes the behavior of the
view call, to always use the current size of ssim_value, avoiding this
issue but keeping the behavior to concatenate in a single row vector.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [x] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [x] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

Signed-off-by: Jakob Weigand <jakob.weigand@tum.de>