added spacing to surface distances calculations (#6144)
Fixes #6137 .

### Description

A user can now pass a `spacing` parameter to surface distance metrics
(Hausdorff distance, surface distance, surface dice), so that correct
results can be obtained for images with non-isotropic spacing (e.g.
(0.5, 0.5, 2) mm).
If `spacing` is a sequence, it must be of length equal to the image
dimensions; if a single number, this spacing is used for all axes. If
``None``, spacing of unity is used. Defaults to ``None``so that current
behaviour is preserved.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [x] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [x] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: gasperp <gasper.podobnik@gmail.com>