Refactor multi-node running command into dedicated functions (#6623)
Fixes #6567  .

### Description

The PR aims to refactor the multi-node command preparation and running
for auto3dseg.
In the initial draft, I assume the functions are only for Auto3DSeg and
they are internal, so I put them in the Auto3DSeg `utils`. I am open to
changes if we think the usage can be more general.

Some details:
To address the 3 variations of commands used in Auto3DSeg:
1. `python script.py <options>`
2. `torchrun <specs> script.py <options>` 
3. `bcprun <specs> -c python script.py <options>`

I split the `<options>` and `<specs>` passing in different stages
- `<options>` is in preparation stage, e.g. `create_cmd` in Auto3DSeg
- `<specs>` is related to device configuration, and so it is in the
launch stage e.g. `run_cmd` in Auto3DSeg

Each variation has its version of preparation and launching.



### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: Mingxin <18563433+mingxin-zheng@users.noreply.github.com>