Add loss_fn to IgniteMetric and rename to IgniteMetricHandler (#6695)
### Description

As explained in #6693 I would like to use the DiceCELoss as a train
metric as well. This branch adds a very crude but working version of
that.
The added tests, which I copied from the MeanDice metric, do still fail.
It would be cool if someone more experienced could check, what needs to
be done there.
I ran the code with my full DeepEdit setup and it appears to be working
just fine there.

No formatting checks done so far - I want to find out first if this code
is useful for others. Docs not updated yet either.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [x] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: Matthias Hadlich <matthiashadlich@posteo.de>