6699 random sort grid patch (#6701)
### Description

The PR introduces a random sorting option to the `GridPatch,
RandGridPatch, GridPatchd, RandGridPatchd` transforms. So far, the
transforms allowed patches to be sorted according to the sum of their
intensity or in their default order (i.e., sliding-window manner). With
this PR, random extraction of patches as described in [Myronenko et
al.](https://arxiv.org/abs/2111.01556) is enabled.

### Types of changes
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [x] New tests added to cover the changes.

---------

Signed-off-by: jmnolte <jakob.nolte@web.de>
Signed-off-by: Jakob Nolte <94493465+jmnolte@users.noreply.github.com>