Add ultrasound confidence map to transforms (#6709)
### Description

This transform uses the method introduced by Karamalis et al. in
https://doi.org/10.1016/j.media.2012.07.005 to
compute confidence maps on ultrasound images. 

### Possible Problems

- I am not entirely sure if the "transforms" section is the right place
for this method but I found it the most suitable
since it is not "deep learning" and it is "pre-processing" in a way.

- Current version of the implementation requires GNU Octave to be
installed and defined in the path. This is an odd
dependency, I am aware of that, yet using SciPy does not provide
satisfactory results in terms of speed. If this kind of
dependency is not suitable, I also have a pure SciPy implementation, yet
it runs about x15 slower, and it is slow to work
in real-time, I am open to any feedback.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [x] New tests added to cover the changes.
- [x] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [x] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: Vahit Bugra YESILKAYNAK <bugrayesilkaynak@gmail.com>