SwinUNETR refactored img_size parameter and removed checkpointing dep… (#7093)
### Description
Make two changes for the SwinUNETR network:
- The image_size parameter does not seem to have an effect apart from
checking shape compatibility in the beginning. This is now expressed in
the docstring and the parameter will be deprecated in the future.
Instead the shape compatibility is checked during the forward pass on
the actual shape
- newer pytorch versions accept a parameter
[use_reentrant](https://pytorch.org/docs/2.1/checkpoint.html). The old
default of True is deprecated in favor of True. This PR sets the
parameter to true and therefore adopts the recommended value and removes
the warning.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: John Zielke <john.zielke@snkeos.com>