Fixing Numpy requirements to exclude 2.0 (#7859)
### Description

This fixes the version of Numpy to be less than 2.0 so that we can get a
working setup guaranteed. Numpy 2.0 isn't supported yet by MONAI and
will need a number of changes to be compatible.

This is only sometimes a problem because a number of dependencies we
have require Numpy<2.0. When these are absent however, such as with
minimal installs, Numpy 2.0 can be installed and causes problems. This
is a temporary fix until we have worked out an update that's compatible
with versions before and after 2.0.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

Signed-off-by: Eric Kerfoot <eric.kerfoot@kcl.ac.uk>